requirements.txt
asgiref==3.8.1
attrs==24.2.0
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
environ==1.0
importlib_resources==6.4.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
packaging==24.1
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==74.0.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
